# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_bitbiz_hias-core_domains_policy_schema.BulkRenewalRequest:
    properties:
      policy_ids:
        items:
          type: string
        type: array
    required:
    - policy_ids
    type: object
...
  github_com_bitbiz_hias-core_domains_policy_schema.InitiateRenewalRequest:
    properties:
      expires_at:
        type: string
      new_plan_id:
        type: string
      policy_id:
        type: string
      renewal_date:
        type: string
    required:
    - policy_id
    - renewal_date
    type: object
...
  github_com_bitbiz_hias-core_domains_policy_schema.RejectRenewalRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Health Insurance Administration System - Core API
  title: HIAS Core API
  version: "1.0"
paths:
...
  /api/v1/policies/{id}/renewals:
    get:
      consumes:
      - application/json
      description: List all renewals associated with a policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List renewals for a policy
      tags:
      - Renewals
    post:
      consumes:
      - application/json
      description: Initiate a renewal process for a policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: Renewal details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.InitiateRenewalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initiate a policy renewal
      tags:
      - Renewals
...
  /api/v1/renewals/{id}:
    get:
      consumes:
      - application/json
      description: Get renewal details by ID
      parameters:
      - description: Renewal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a renewal
      tags:
      - Renewals
  /api/v1/renewals/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve a pending renewal by ID
      parameters:
      - description: Renewal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve a renewal
      tags:
      - Renewals
  /api/v1/renewals/{id}/complete:
    post:
      consumes:
      - application/json
      description: Complete an approved renewal process
      parameters:
      - description: Renewal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Complete a renewal
      tags:
      - Renewals
  /api/v1/renewals/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject a pending renewal by ID with a reason
      parameters:
      - description: Renewal ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.RejectRenewalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject a renewal
      tags:
      - Renewals
  /api/v1/renewals/bulk:
    post:
      consumes:
      - application/json
      description: Initiate renewal processes for multiple policies at once
      parameters:
      - description: List of policy IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_bitbiz_hias-core_domains_policy_schema.BulkRenewalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk initiate renewals
      tags:
      - Renewals
  /api/v1/renewals/expire:
    post:
      consumes:
      - application/json
      description: Expire all pending renewals that have passed their deadline
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Expire pending renewals
      tags:
      - Renewals
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"